Redesigned the Sign-In and Sign-Up pages - changed color theme and added redirection, 'remember-me'  buttons
