+++ qala_frontend/package.json
@@ -13,7 +13,8 @@
     "react-dom": "^19.0.0",
     "react-leaflet": "^5.0.0",
     "react-router-dom": "^7.1.1",
-    "react-scripts": "5.0.1"
+    "react-scripts": "5.0.1",
+    "styled-elements": "^1.0.14"
   },
   "scripts": {
     "start": "react-scripts start",
