+++ qala_frontend/src/App.jsx
@@ -7,12 +7,17 @@ import LoginPage from "./routes/Login";
 import { Box } from "@mui/material";
 import { AuthProvider } from "./context/AuthProvider";
 import AccountDashboard from "./routes/Profile";
+import Events from "./routes/Events";
 
 function App() {
   const routes = useRoutes([
     {
       path: "/",
       element: <Home />,
+    },
+    {
+      path: "/events",
+      element: <Events />,
     },
     {
       path: "/map",
